Add more tampere brand colors for different elements
We need to overwrite some scss variables to get the
result we want. In order to do that, we need to import
the variables in a specific order.

Some of the styles already applied can be removed due to
the fact that the changed variables will do the styling
for us.